Move auth tests to beholder package, unexport new auth
